Broke functions down into appropriate modules
