+++ bluemira/neutronics/import_diagram.txt
@@ -1,0 +1,16 @@
+quick_tbr_heating
+        | <-------------------- params.py
+        |                       |     |
+        | <-- constants.py      |     |
+        |           |           |     |
+        |           V           V     |
+        | <-- result_presentation.py  |
+        |                             V
+        | <------------------------ volume_functions.py
+        |
+        | <-- make_materials.py <-- materials_definition.py
+        |
+        | <-- make_geometry.py
+        |
+        V
+< end of program >
