+++ bluemira/neutronics/make_materials.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import copy
 import dataclasses
+import os
 from dataclasses import dataclass
 from enum import Enum, auto
 from typing import Union
